Refactor Application state code
