+++ components/LearnMoreModal/index.js
@@ -1,11 +1,8 @@
-import {
-  useGetInvitedModalToggle,
-  useLearnMoreModalOpen,
-  useLearnMoreModalToggle,
-} from "states/Application";
+import { useGetInvitedModalToggle } from "components/GetInvitedModal";
 import { useRouter } from "next/router";
 import { Solo, Together } from "./benefits";
 import Modal, { CloseButton } from "./modal";
+import { useLearnMoreModalOpen, useLearnMoreModalToggle } from "./state";
 
 const LearnMoreModal = () => {
   const open = useLearnMoreModalOpen();
@@ -54,3 +51,5 @@ const LearnMoreModal = () => {
 };
 
 export default LearnMoreModal;
+
+export { useLearnMoreModalOpen, useLearnMoreModalToggle };
