--- states/Application/index.js
@@ -1,12 +1,0 @@
-export {
-  useGetInvitedModalOpen,
-  useGetInvitedModalToggle,
-  useLearnMoreModalOpen,
-  useLearnMoreModalToggle,
-  useToggleCreateModal,
-  useToggleSignInModal,
-  useUpdateWalletModalView,
-  useWalletModalOpen,
-  useWalletModalToggle,
-  useWalletModalView,
-} from "./hooks";
