--- states/Application/states.js
@@ -1,9 +1,0 @@
-import { newRidgeState } from "react-ridge-state";
-
-export const walletModalViewState = newRidgeState("CREATE");
-
-export const walletModalState = newRidgeState(false);
-
-export const getInvitedModalState = newRidgeState(false);
-
-export const learnMoreModalState = newRidgeState(false);
